Add attribute on monobehaviour types
